Remove unused ResourceMap from kube
